requirements.txt
# --- FastAPI Core ---
fastapi==0.115.5
uvicorn[standard]==0.30.6

# --- Environment & Settings ---
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# --- Database (SQL Server via ODBC) ---
pyodbc==5.1.0

# --- PDF Parsing & Generation ---
pdfplumber==0.11.4
pdfminer.six==20231228
pymupdf==1.24.14
pdfkit==1.0.0      # Requires wkhtmltopdf installed on server

# --- Web / Templates ---
jinja2==3.1.4
aiofiles==24.1.0
python-multipart==0.0.9

# --- Performance / Utilities ---
orjson==3.10.7
charset-normalizer==3.4.0
pandas==2.2.3
